+++ src/blocks/download/edit.tsx
@@ -7,7 +7,9 @@ import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
 import { RecoilRoot, useRecoilState } from 'recoil'
-import { Box, Group, SetupEditorEnv, SourceSelect, Text } from '../../components'
+import { Box, Group, Text } from '../../components'
+import { SourceSelect } from '../../components/SourceSelect'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import json from './block.json'
 import { Attributes } from './index'
 import { wizardState } from './state'
