+++ src/blocks/download/hydration.tsx
@@ -2,7 +2,7 @@ import { InseriRoot } from '@inseri/lighthouse'
 import domReady from '@wordpress/dom-ready'
 import { createRoot } from '@wordpress/element'
 import { RecoilRoot } from 'recoil'
-import { InseriThemeProvider } from '../../components'
+import { InseriThemeProvider } from '../../components/InseriThemeProvider'
 import { generateQuerySelector } from '../../utils'
 import config from './block.json'
 import View from './view'
