+++ src/blocks/export/edit.tsx
@@ -5,7 +5,8 @@ import { PanelBody, PanelRow } from '@wordpress/components'
 import { select } from '@wordpress/data'
 import { useEffect } from '@wordpress/element'
 import { disablePostExport, enablePostExport } from '../../ApiServer'
-import { SetupEditorEnv, TextInput } from '../../components'
+import { TextInput } from '../../components'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import config from './block.json'
 import { Attributes } from './index'
