+++ src/blocks/html/edit.tsx
@@ -6,7 +6,9 @@ import { PanelBody, PanelRow, TextControl, ToolbarButton, ToolbarGroup } from '@
 import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Group, SetupEditorEnv, SourceSelect, Text } from '../../components'
+import { Box, Group, Text } from '../../components'
+import { SourceSelect } from '../../components/SourceSelect'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import json from './block.json'
 import { Attributes } from './index'
