+++ src/blocks/localFileImport/edit.tsx
@@ -3,7 +3,8 @@ import { InspectorControls } from '@wordpress/block-editor'
 import type { BlockEditProps } from '@wordpress/blocks'
 import { PanelBody, PanelRow, ResizableBox, TextControl, TextareaControl, ToggleControl } from '@wordpress/components'
 import { __ } from '@wordpress/i18n'
-import { MultiSelect, SetupEditorEnv, createStyles } from '../../components'
+import { MultiSelect, createStyles } from '../../components'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import { COMMON_CONTENT_TYPES } from '../../utils'
 import { default as config, default as json } from './block.json'
