+++ src/blocks/localFileImport/hydration.tsx
@@ -2,7 +2,7 @@ import { InseriRoot } from '@inseri/lighthouse'
 import domReady from '@wordpress/dom-ready'
 import { createRoot } from '@wordpress/element'
 import { generateQuerySelector } from '../../utils'
-import { InseriThemeProvider } from '../../components'
+import { InseriThemeProvider } from '../../components/InseriThemeProvider'
 import { StateProvider } from '../../components/StateProvider'
 import config from './block.json'
 import { storeCreator } from './state'
