+++ src/blocks/plotly/edit.tsx
@@ -7,7 +7,9 @@ import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
 import { useMap } from 'react-use'
-import { Accordion, ActionIcon, Box, Button, Group, Select, SetupEditorEnv, SourceSelect, Stack, Text } from '../../components'
+import { Accordion, ActionIcon, Box, Button, Group, Select, Stack, Text } from '../../components'
+import { SourceSelect } from '../../components/SourceSelect'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import json from './block.json'
 import { Attributes } from './index'
