+++ src/blocks/rCode/edit.tsx
@@ -6,7 +6,9 @@ import { PanelBody, PanelRow, ResizableBox, TextControl, ToggleControl, ToolbarB
 import { useEffect, useState } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Button, Group, SetupEditorEnv, SourceSelect, Text } from '../../components'
+import { Box, Button, Group, Text } from '../../components'
+import { SourceSelect } from '../../components/SourceSelect'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import { ExtendedView } from '../../components/ExtendedView'
 import { HidingWrapper } from '../../utils'
