+++ src/blocks/share/edit.tsx
@@ -3,7 +3,8 @@ import { InspectorControls } from '@wordpress/block-editor'
 import type { BlockEditProps } from '@wordpress/blocks'
 import { PanelBody, PanelRow } from '@wordpress/components'
 import { select } from '@wordpress/data'
-import { Box, Group, SetupEditorEnv, Switch, TextInput } from '../../components'
+import { Box, Group, Switch, TextInput } from '../../components'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import config from './block.json'
 import { Attributes } from './index'
