+++ src/blocks/slider/edit.tsx
@@ -2,7 +2,8 @@ import { InseriRoot } from '@inseri/lighthouse'
 import { InspectorControls } from '@wordpress/block-editor'
 import type { BlockEditProps } from '@wordpress/blocks'
 import { PanelBody, PanelRow } from '@wordpress/components'
-import { Box, Group, NumberInput, SegmentedControl, SetupEditorEnv, Stack, Switch, TextInput, Title } from '../../components'
+import { Box, Group, NumberInput, SegmentedControl, Stack, Switch, TextInput, Title } from '../../components'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import config from './block.json'
 import { Attributes } from './index'
