+++ src/blocks/textEditor/edit.tsx
@@ -6,7 +6,8 @@ import { PanelBody, PanelRow, ResizableBox, TextControl, ToggleControl, ToolbarB
 import { useEffect } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { edit } from '@wordpress/icons'
-import { Box, Group, Select, SetupEditorEnv, Text } from '../../components'
+import { Box, Group, Select, Text } from '../../components'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import { HidingWrapper, TEXTUAL_CONTENT_TYPES } from '../../utils'
 import config from './block.json'
