+++ src/blocks/webApi/edit.tsx
@@ -7,7 +7,9 @@ import { PanelBody, PanelRow, TextControl, ToggleControl, ToolbarButton, Toolbar
 import { useEffect } from '@wordpress/element'
 import { __ } from '@wordpress/i18n'
 import { useMap } from 'react-use'
-import { Box, Button, Group, Modal, Select, SetupEditorEnv, SourceSelect, Stack, Text, TextInput, createStyles, getStylesRef } from '../../components'
+import { Box, Button, Group, Modal, Select, Stack, Text, TextInput, createStyles, getStylesRef } from '../../components'
+import { SourceSelect } from '../../components/SourceSelect'
+import { SetupEditorEnv } from '../../components/SetupEditorEnv'
 import { StateProvider, useGlobalState } from '../../components/StateProvider'
 import { ContentTypeSelect } from '../../components/ContentTypeSelect'
 import { DetailViewBody } from '../../components/DetailViewBody'
