+++ src/components/ExtendedView.tsx
@@ -6,7 +6,8 @@ import { __ } from '@wordpress/i18n'
 import type { AllotmentHandle } from 'allotment'
 import { Allotment } from 'allotment'
 import { COMMON_CONTENT_TYPES, Z_INDEX_ABOVE_ADMIN, isVariableValid } from '../utils'
-import { ActionIcon, Box, Button, CodeEditor, Group, Modal, Stack, Text, TextInput, createStyles } from './'
+import { ActionIcon, Box, Button, Group, Modal, Stack, Text, TextInput, createStyles } from './'
+import { CodeEditor } from './CodeEditor'
 import { useGlobalState } from './StateProvider'
 import { SelectWithAction, SourceSelectWithAction } from './SelectWithAction'
 import { TopBar } from './TopBar'
