+++ src/components/index.tsx
@@ -1,6 +1,1 @@
 export * from '@mantine/core'
-export * from './SortableTh'
-export * from './SourceSelect'
-export * from './CodeEditor'
-export * from './InseriThemeProvider'
-export * from './SetupEditorEnv'
